// src/dtos/referenceAttribute.dto.ts
/* eslint-disable */

export class refAttrDto {
    tenantId: number
    attributeName: string
    displayName: string
    attributeType : string
    referenceMasterId : string
    status : boolean

}

export class refAttrByIdDto {
    id: number
    tenantId: number
    attributeName: string
    displayName: string
    attributeType : string
    referenceMasterId : string
    status : boolean
}